WebApp Basketball Club System using ASP.net, C_Sharp, SQL server, HTML, CSS, Js, jQuery and Bootstrap
WebApp Basketball Club System using ASP.net, C_Sharp, SQL server, HTML, CSS, Js, jQuery and Bootstrap.

promote the club by showing some information and prompt the activities as well as ability to reserve some basketball equipment.
the user can register to the system and log in and also can use remember me feature. there is a message that will appear as feedback for the user.

the manager can log in to the system and manage users. editing information, update pictures, update the role, delete the user and update data as well as add a new user and assign a role for the user such as user, manager, coach.

there are many functions that have been implemented in this project, you can check them out.